pallet-remarks: Move handler back to lib
